+++ .travis.yml
@@ -1,0 +1,14 @@
+language: node_js
+dist: trusty
+sudo: required
+node_js: 6
+install:
+  - npm install -g bower
+  - npm install && bower install
+script:
+  - npm run build:browser
+after_success:
+- >-
+  test $TRAVIS_TAG &&
+  echo $GITHUB_TOKEN | pulp login &&
+  echo y | pulp publish --no-push
