+++ example/index.html
@@ -12,10 +12,6 @@
   </style>
   <body>
     <div id="map"></div>
-    <div id="heatmap-container">
-      <canvas id="heatmap" width="1000" height="600">
-      </canvas>
-    </div>
     <div id="heatmap-leaflet"></div>
   </body>
 </html>
