--- example/src/Heatmap.js
@@ -1,1 +1,0 @@
-trololo
