+++ src/Leaflet/Plugin/Heatmap/Internal/Canvas.purs
@@ -2,6 +2,7 @@ module Leaflet.Plugin.Heatmap.Internal.Canvas
   ( draw
   , createCanvas
   , canvasToElement
+  , defaultOptions
   , HeatmapOptions
   , HeatmapPoint
   , module G
